fix(go): finish the level-type change, and stop synthesizing a pointer from a cgo.Handle
Two things, both surfaced by finally getting a working local Go
typecheck.

The `[]uint8` -> `[]uint32` change missed a literal in
`subnet_golden_vectors_test.go` and a doc example in `subnet.go`. CI's
`go test` caught the literal; I could not, because `go vet` on this host
was silently compiling the package WITHOUT its cgo files and reporting
`undefined: MeshNode` instead of anything useful. Root cause was
`CGO_ENABLED=0` plus no `gcc` on PATH. With `CGO_ENABLED=1` and msys64's
mingw gcc, vet typechecks the whole package — test files included — and
would have caught this before the push.

That same working vet then reported one genuine pre-existing finding:
`unsafe.Pointer(uintptr(cgoHandle))` in `meshos.go`. A `cgo.Handle` is an
integer TOKEN, not a pointer — the substrate never dereferences
`user_ctx`, it only hands the value back to a trampoline that converts it
straight to a Handle again. But synthesizing an `unsafe.Pointer` from an
integer is invalid under Go's unsafe.Pointer rules, and under
`-d=checkptr` (which `go test -race` enables) it can fault, so this was a
latent race-mode crash rather than a style nit.

The cast moves into C, where it is well-defined for a value that is never
dereferenced: a `static inline void* topGoMeshOsCtxFromHandle(uintptr_t)`
in the cgo preamble. The reverse direction stays in Go —
`cgo.Handle(uintptr(userCtx))` is a pointer->integer conversion, which is
legal and unflagged.

`CGO_ENABLED=1 go vet .` is now clean across the entire package, with no
findings at all. `gofmt` clean.

+++ go/subnet_golden_vectors_test.go
@@ -176,7 +176,7 @@ func TestSubnetNamedExportMarshalsForTheRustConstructor(t *testing.T) {
 		Name:   "factory-export",
 		Access: SubnetAccessGranted,
 		Binding: SubnetExportBinding{
-			Subnet:        SubnetRef{AuthorityHex: hex64(), Path: SubnetPath{Levels: []uint8{3, 9}}},
+			Subnet:        SubnetRef{AuthorityHex: hex64(), Path: SubnetPath{Levels: []uint32{3, 9}}},
 			TopologyEpoch: 7,
 		},
 	}
